requirements.txt
python-dotenv==1.1.0
fastmcp==2.7.0
openai==1.75.0
tenacity==9.1.2
colorlog==6.9.0
tiktoken==0.9.0 
colorama==0.4.6 
loguru==0.7.3
retry==0.9.2 
pandas==2.2.3
yt_dlp
opencv-python==4.11.0.86
scenedetect==0.6.6 
assemblyai==0.40.2  
python-pptx==1.0.2  
docx2markdown==0.1.1    
chunkr_ai
xmltodict==0.14.2
crawl4ai
PyGithub
matplotlib
geopy
requests
psutil
fpdf2
